Refactored skills
