+++ src/stp/new_tactics/AvoidBall.cpp
@@ -7,8 +7,8 @@
 #include <roboteam_utils/Circle.h>
 #include <roboteam_utils/Tube.h>
 
-#include "stp/new_skills/GoToPos.h"
-#include "stp/new_skills/Rotate.h"
+#include "stp/skills/GoToPos.h"
+#include "stp/skills/Rotate.h"
 #include "utilities/GameStateManager.hpp"
 
 namespace rtt::ai::stp::tactic {
