+++ src/stp/new_tactics/Formation.cpp
@@ -4,8 +4,8 @@
 
 #include "stp/new_tactics/Formation.h"
 
-#include "stp/new_skills/GoToPos.h"
-#include "stp/new_skills/Rotate.h"
+#include "stp/skills/GoToPos.h"
+#include "stp/skills/Rotate.h"
 
 namespace rtt::ai::stp::tactic {
