+++ src/stp/new_tactics/GetBall.cpp
@@ -7,8 +7,8 @@
 #include <utilities/GameStateManager.hpp>
 
 #include "control/ControlUtils.h"
-#include "stp/new_skills/GoToPos.h"
-#include "stp/new_skills/Rotate.h"
+#include "stp/skills/GoToPos.h"
+#include "stp/skills/Rotate.h"
 #include "world/FieldComputations.h"
 
 namespace rtt::ai::stp::tactic {
