+++ src/stp/new_tactics/GetBallInDirection.cpp
@@ -3,8 +3,8 @@
 //
 
 #include "stp/new_tactics/GetBallInDirection.h"
-#include "stp/new_skills/GoToPos.h"
-#include "stp/new_skills/Rotate.h"
+#include "stp/skills/GoToPos.h"
+#include "stp/skills/Rotate.h"
 
 namespace rtt::ai::stp::tactic {
     GetBallInDirection::GetBallInDirection() {
