+++ src/stp/new_tactics/Halt.cpp
@@ -4,7 +4,7 @@
 
 #include "stp/new_tactics/Halt.h"
 
-#include "stp/new_skills/Rotate.h"
+#include "stp/skills/Rotate.h"
 
 namespace rtt::ai::stp::tactic {
